Issue 155: SAX parsing and 1 thread per XML file
